Move MaliciousSiteProtectionPreferences to its own file
